Literals now backed by BigInteger to support larger values

+++ compiler/src/main/kotlin/com/uabutler/ast/node/TokenNodes.kt
@@ -1,9 +1,11 @@
 package com.uabutler.ast.node
 
+import java.math.BigInteger
+
 data class IdentifierNode(
     val value: String,
 ): PersistentNode
 
 data class IntegerLiteralNode(
-    val value: Int,
+    val value: BigInteger,
 ): PersistentNode
